// src/app/_dto/payrollsetupdetail.dto.ts
export class payrollsetupdetail {
    DPayroll_Grid_SysID: number;
    DPayroll_SlNo: number;
    DPayroll_Add_SysID: number;
    DPayroll_Add_Code: string;
    DPayroll_Add_Name: string;
    DPayroll_AccAdd_SysID: number;
    DPayroll_AccAdd_Code: string;
    DPayroll_AccAdd_Name: string;
    DPayroll_Ded_SysID: number;
    DPayroll_Ded_Code: string;
    DPayroll_Ded_Name: string;
    DPayroll_AccDed_SysID: number;
    DPayroll_AccDed_Code: string;
    DPayroll_AccDed_Name: string;
    DPayroll_Remark_SysID: number;
    DPayroll_Remark_Code: string;
    DPayroll_Remark_Name: string;
    selected :boolean;
}